Version 2.1.0 (#38)

+++ tests/web_tests/test_assert_screenshot.py
@@ -5,14 +5,14 @@ import pytest_rerunfailures
 
 
 @pytest.mark.parametrize('with_name', [True, False], ids=['screenshot name given', 'screenshot name missed'])
-def test_screenshot(base_playground_page, driver_engine, driver_name, platform, with_name):
-    filename = f'{driver_engine}-{driver_name}-{platform}-kube' if with_name else ''
+def test_screenshot(base_playground_page, driver_name, platform, with_name):
+    filename = f'{driver_name}-{platform}-kube' if with_name else ''
     base_playground_page.kube.scroll_into_view().assert_screenshot(filename)
 
 
 @pytest.mark.parametrize('with_name', [True, False], ids=['screenshot name given', 'screenshot name missed'])
-def test_screenshot_name_with_suffix(base_playground_page, driver_engine, driver_name, platform, with_name):
-    filename = f'{driver_engine}-{driver_name}-{platform}-kube' if with_name else ''
+def test_screenshot_name_with_suffix(base_playground_page, driver_name, platform, with_name):
+    filename = f'{driver_name}-{platform}-kube' if with_name else ''
     base_playground_page.kube.scroll_into_view().assert_screenshot(filename, name_suffix='first')
     base_playground_page.kube.scroll_into_view().assert_screenshot(filename, name_suffix='second')
 
@@ -29,7 +29,8 @@ def file(request):
     filename = 'reference_with_rerun'
     yield filename
     request.node.session.config.option.reruns = 0
-    os.remove(f'{os.getcwd()}/tests/adata/visual/reference/{filename}.png')
+    if not request.config.option.sv:
+        os.remove(f'{os.getcwd()}/tests/adata/visual/reference/{filename}.png')
 
 
 def test_screenshot_without_reference_and_rerun(base_playground_page, file, request):
@@ -39,7 +40,8 @@ def test_screenshot_without_reference_and_rerun(base_playground_page, file, requ
     except AssertionError:
         pass
     else:
-        if not request.config.getoption('--gr') or not request.config.getoption('--hgr'):
+        options = request.config.option
+        if not (options.gr or options.hgr or options.sv):
             raise Exception('Unexpected behavior')
